constructors do not generate signals on their own

+++ source/EngineGpuKernels/SignalProcessor.cuh
@@ -16,7 +16,7 @@ public:
     __inline__ __device__ static void createEmptySignal(Cell* cell);
     __inline__ __device__ static float2 calcReferenceDirection(SimulationData& data, Cell* cell);
 
-    __inline__ __device__ static bool isAutoTriggered(SimulationData& data, Cell* cell, int autoTriggerInterval);
+    __inline__ __device__ static bool isAutoTriggered(SimulationData& data, Cell* cell, uint8_t autoTriggerInterval);
     __inline__ __device__ static bool isTriggeredAndCreateSignalIfTriggered(SimulationData& data, Cell* cell, uint8_t autoTriggerInterval);
 };
 
@@ -163,7 +163,7 @@ __inline__ __device__ float2 SignalProcessor::calcReferenceDirection(SimulationD
     return Math::normalized(data.cellMap.getCorrectedDirection(cell->pos - cell->connections[0].cell->pos));
 }
 
-__inline__ __device__ bool SignalProcessor::isAutoTriggered(SimulationData& data, Cell* cell, int autoTriggerInterval)
+__inline__ __device__ bool SignalProcessor::isAutoTriggered(SimulationData& data, Cell* cell, uint8_t autoTriggerInterval)
 {
     auto triggerInterval = max(MAX_SIGNAL_RELAXATION_TIME + 1, autoTriggerInterval);
     return (data.timestep + cell->creatureId) % triggerInterval == 0;
